+++ README.txt
@@ -1,0 +1,5 @@
+GitHub Commands:
+
+    *Commit - Save Local Changes
+    *Push - Publish Changes to Branch
+    *Update - Retrieve Files from selected branch
