--- RF_Integration_Main/README.txt
@@ -1,5 +1,0 @@
-GitHub Commands:
-
-    *Commit - Save Local Changes
-    *Push - Publish Changes to Branch
-    *Update - Retrieve Files from selected branch
