+++ __pycache__/CSVAlgoFoxtrot.cpython-310.pyc
